# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      operationId: getProject
      tags: [projects]
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    patch:
      operationId: updateProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdate" }
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    delete:
      operationId: deleteProject
      tags: [projects]
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/tasks:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: integer }
    get:
      operationId: listTasks
      tags: [tasks]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskInput" }
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    patch:
      operationId: updateTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskUpdate" }
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
    delete:
      operationId: deleteTask
      tags: [tasks]
      responses:
        "204":
          description: Deleted

  # ---------- Calendar ----------
...
  /system-state:
    get:
      operationId: listSystemState
      tags: [system-state]
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Daily system state history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SystemStateDaily" }
...
  /system-state/warnings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    patch:
      operationId: dismissStateWarning
      tags: [system-state]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StateWarningUpdate" }
      responses:
        "200":
          description: Updated warning
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateWarning" }
...
  /ultra-score/calculate:
    post:
      operationId: calculateUltraScore
      tags: [system-state]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UltraScoreCalculateInput" }
      responses:
        "200":
          description: Calculated ultra score result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemStateDaily" }

  # ---------- Notifications ----------
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, userId, title]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        hubId: { type: ["integer", "null"] }
        title: { type: string }
        status: { type: string }
        priority: { type: ["string", "null"] }
        sprint: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"], format: date }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProjectInput:
      type: object
      required: [title]
      properties:
        tenantId: { type: string, format: uuid }
        hubId: { type: integer }
        title: { type: string, minLength: 1 }
        status: { type: string }
        priority: { type: string }
        sprint: { type: string }
        notes: { type: string }
        dueDate: { type: string, format: date }
    ProjectUpdate:
      type: object
      properties:
        title: { type: string }
        status: { type: string }
        priority: { type: string }
        sprint: { type: string }
        notes: { type: string }
        dueDate: { type: string, format: date }
    Task:
      type: object
      required: [id, projectId, title]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: ["string", "null"] }
        importance: { type: ["integer", "null"] }
        dueDate: { type: ["string", "null"], format: date }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TaskInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        importance: { type: integer }
        dueDate: { type: string, format: date }
    TaskUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        importance: { type: integer }
        dueDate: { type: string, format: date }

    # ---------- Calendar ----------
...
    SystemStateDaily:
      type: object
      required: [id, userId, stateDate, ultraScore, state]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        stateDate: { type: string, format: date }
        ultraScore: { type: number }
        state: { type: string }
        priorityZone: { type: ["string", "null"] }
        strongestHubId: { type: ["integer", "null"] }
        weakestHubId: { type: ["integer", "null"] }
        stateReasons: {}
        createdAt: { type: string, format: date-time }
    StateWarning:
      type: object
      required: [id, userId, warningType, warningText]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        warningType: { type: string }
        warningText: { type: string }
        severity: { type: ["string", "null"] }
        relatedHubId: { type: ["integer", "null"] }
        relatedProjectId: { type: ["integer", "null"] }
        relatedHabitId: { type: ["integer", "null"] }
        dismissed: { type: boolean }
        dismissedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    StateWarningUpdate:
      type: object
      properties:
        dismissed: { type: boolean }
    UltraScoreCalculateInput:
      type: object
      properties:
        tenantId: { type: string, format: uuid }
        stateDate: { type: string, format: date }
...